Refactor auth routes into dedicated blueprint
