+++ website/blueprints/auth/routes.py
@@ -1,0 +1,47 @@
+from functools import wraps
+from flask import (
+    Blueprint,
+    render_template,
+    request,
+    redirect,
+    url_for,
+    session,
+    flash,
+)
+
+from ...utils.allowlist import verify_user
+
+auth_bp = Blueprint("auth", __name__)
+
+
+def login_required(view):
+    @wraps(view)
+    def wrapped(*args, **kwargs):
+        if "user" not in session:
+            return redirect(url_for("auth.login"))
+        return view(*args, **kwargs)
+
+    return wrapped
+
+
+@auth_bp.route("/login", methods=["GET", "POST"])
+def login():
+    if request.method == "POST":
+        email = request.form.get("email", "").strip().lower()
+        password = request.form.get("password", "")
+        if verify_user(email, password):
+            session["user"] = email
+            return redirect(url_for("core.generador"))
+        flash("Credenciales inválidas", "warning")
+    return render_template("login.html")
+
+
+@auth_bp.route("/logout")
+def logout():
+    session.pop("user", None)
+    return redirect(url_for("auth.login"))
+
+
+@auth_bp.route("/register")
+def register():
+    return redirect(url_for("auth.login"))
