+++ website/blueprints/core.py
@@ -1,59 +1,9 @@
-from functools import wraps
-from flask import (
-    Blueprint,
-    render_template,
-    request,
-    redirect,
-    url_for,
-    session,
-    flash,
-)
+from flask import Blueprint, render_template
 from flask_wtf.csrf import CSRFError
 
-from ..utils.allowlist import verify_user
+from .auth.routes import login_required
 
 bp = Blueprint("core", __name__)
-
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-def login_required(view):
-    @wraps(view)
-    def wrapped(*args, **kwargs):
-        if "user" not in session:
-            return redirect(url_for("core.login"))
-        return view(*args, **kwargs)
-
-    return wrapped
-
-
-# ---------------------------------------------------------------------------
-# Routes
-# ---------------------------------------------------------------------------
-
-@bp.route("/login", methods=["GET", "POST"])
-def login():
-    if request.method == "POST":
-        email = request.form.get("email", "").strip().lower()
-        password = request.form.get("password", "")
-        if verify_user(email, password):
-            session["user"] = email
-            return redirect(url_for("core.generador"))
-        flash("Credenciales inválidas", "warning")
-    return render_template("login.html")
-
-
-@bp.route("/logout")
-def logout():
-    session.pop("user", None)
-    return redirect(url_for("core.login"))
-
-
-@bp.route("/register")
-def register():
-    return redirect(url_for("core.login"))
 
 
 @bp.route("/generador")
@@ -73,10 +23,6 @@ def resultados():
 def configuracion():
     return render_template("configuracion.html")
 
-
-# ---------------------------------------------------------------------------
-# Error handlers
-# ---------------------------------------------------------------------------
 
 @bp.app_errorhandler(CSRFError)
 def handle_csrf_error(error):
