+++ website/templates/landing.html
@@ -39,7 +39,7 @@
         <li class="nav-item"><a class="nav-link px-3" href="#pricing">Precios</a></li>
         <li class="nav-item"><a class="nav-link px-3" href="#faq">FAQ</a></li>
         <li class="nav-item"><a class="nav-link px-3" href="{{ url_for('contacto') }}">Contacto</a></li>
-        <li class="nav-item ms-lg-3"><a class="btn btn-outline-secondary" href="{{ url_for('core.login') }}">Iniciar sesión</a></li>
+        <li class="nav-item ms-lg-3"><a class="btn btn-outline-secondary" href="{{ url_for('auth.login') }}">Iniciar sesión</a></li>
       </ul>
     </div>
   </div>
@@ -56,7 +56,7 @@
           menos exceso, cero déficit y exportación a Excel lista para operación.
         </p>
         <div class="d-flex gap-3">
-          <a href="{{ url_for('core.login') }}" class="btn btn-brand btn-lg px-4">Iniciar sesión</a>
+          <a href="{{ url_for('auth.login') }}" class="btn btn-brand btn-lg px-4">Iniciar sesión</a>
           <a href="#features" class="btn btn-outline-secondary btn-lg px-4">Ver funciones</a>
         </div>
         <div class="d-flex gap-4 mt-4">
@@ -374,7 +374,7 @@
     <div class="d-flex gap-3 small">
       <a class="footer-link" href="#features">Funciones</a>
       <a class="footer-link" href="#pricing">Precios</a>
-      <a class="footer-link" href="{{ url_for('core.login') }}">Iniciar sesión</a>
+      <a class="footer-link" href="{{ url_for('auth.login') }}">Iniciar sesión</a>
     </div>
   </div>
 </footer>
